verify symbols for all .o inside a .a
I verified on the 1981 .o files in libc.a that the symbol lookup table in a .a file specifies
exactly those symbols matching the following conditions for each .o file:
- it's GLOBAL or WEAK (not LOCAL)
- it's defined. i.e., its section number either points to a non NULL section or it's SHN_ABS

One corner case is, some .o file may not have a SYMTAB section. Then they may not have entries in
the symbol lookup table.