figure out the minimal set of .o files needed from a .a
Check the makefile in sar/ for details. sar can accept a list of symbols that are needed, and a list of symbols
defined by other .a files, then it will figure out what is the minimal set of .o files needed for the
target .a file. The list is written to a hardcoded path '/tmp/dep_obj_list'.